requirement.txt
streamlit
fastapi
uvicorn
google-generativeai
python-dotenv
requests
email-validator
python-multipart
typing-extensions